Final Commit
This is a complete working color picker that includes a randomize feature to change the UIView elements in this project. The sliders are configured to change the color values and update them in the first ViewController to display to saved RGB value and display it via the labels on the first UIView.